fix: iOS compile errors and one-way boundary-drawing exit (tc-6he3x.10)
CI compile fixes (PR Gate red on first push, no Swift toolchain in the
cloud session that authored the previous commits to verify against):
- BoundaryDrawingViewModel.swift: add missing `import Combine` so
  `ObservableObject` resolves.
- BoundaryDrawingMapView.swift: convert two `static let`s to computed
  `static var`s -- Swift disallows stored static properties on a
  generic type, which BoundaryDrawingMapView became in the prior commit.
- OnboardingViewModelTests.swift: split the new custom-shape boundary
  drawing coverage into OnboardingViewModelBoundaryDrawingTests.swift --
  the combined file tripped the 400-line file_length lint rule.

Review fix (tc-review-routine flagged 1 high, 1 low; CodeRabbit
independently flagged the same high as a "Major" finding): a user who
purchased the custom-shape upgrade mid-onboarding and tapped Back once
from the drawing step was permanently stranded on a circle zone for the
rest of onboarding -- the upsell card that's the only entry point into
boundaryDrawing hides itself once the tier is already entitled, and
goBack() didn't clear boundaryVertices either. Fixes: goBack() now
clears boundaryVertices when discarding the in-progress shape, and a
new selectCustomShape() plus a "Draw a custom shape instead" affordance
on the radius step gives an already-entitled user a way back in.
tc-review-routine's low finding (vertex-editing logic duplicated
between OnboardingViewModel and WatchZoneEditorViewModel rather than
via a BoundaryDrawingViewModel default implementation) is a
maintainability note, not a functional bug -- left as-is for now.

CodeRabbit's remaining findings (step indicator always showing 5
capsules, missing UI error display for a rejected ring, confirmBoundary
not setting shapeMode, test nitpicks) are left for the PR-comment-
runner's pass per CLAUDE.md policy.

(tc-6he3x.10)
