+++ mobile/ios/packages/town-crier-presentation/Sources/Features/BoundaryDrawingViewModel.swift
@@ -1,3 +1,4 @@
+import Combine
 import TownCrierDomain
 
 /// The minimal surface `BoundaryDrawingMapView` needs from whatever
